Character_RNN_Book_with_Batch_Generator.ipynb
Character based sequence to sequence model implementation using RNN(LSTMs). So for a given book, a sequence of 50 characters is predicted after an example sequence of 100 characters. Batch Generator is used for effective memory utilization so that only 2000 characters are processed at one time and not all of the 770000 characters together.